refactor(shell/sidebar): extract pinboard panel (PR 11/12)
Pull the ~545-LOC pinboard panel out of sidebar/index.js into its
own ES module at shell/js/sidebar/panels/pinboard.js. Follows the
same pattern as the earlier bookmarks + history extractions.

No behaviour change — activateItem still routes 'pinboards' to the
same code path, only via an import now. The tab-context-menu's
"Add to Pinboard" submenu stays in index.js (Task 12) and uses a
new refreshPinboardIfOpen() export instead of reaching into pbState.
The onPinboardItemAdded hook is now installed lazily by the module
on first panel open.
